Create, Save and Reuse Machine Learning Model
Simple demo of how to create, save and reuse classification model. The ability to save and re-use a trained model is an important component in the productionizing of machine learning models.

The model is created in scikit-learn.

The model is saved and re-opened (serialized and deserialized) using pickle.

The data set used for the demo is the Iris Flower Dataset.